move twitch client manager
