+++ src/bot.ts
@@ -3,7 +3,7 @@ import Auth from './net/Auth'
 import ModuleManager from './mod/ModuleManager'
 import WebSocketServer from './net/WebSocketServer'
 import WebServer from './WebServer'
-import TwitchClientManager from './net/TwitchClientManager'
+import TwitchClientManager from './services/TwitchClientManager'
 import ModuleStorage from './mod/ModuleStorage'
 import { logger, setLogLevel } from './common/fn'
 import Users, { User } from './services/Users'
