+++ src/types.ts
@@ -9,7 +9,7 @@ import ModuleManager from './mod/ModuleManager'
 import { GeneralModuleEmotesEventData } from './mod/modules/GeneralModuleCommon'
 import ModuleStorage from './mod/ModuleStorage'
 import Auth from './net/Auth'
-import TwitchClientManager from './net/TwitchClientManager'
+import TwitchClientManager from './services/TwitchClientManager'
 import WebSocketServer, { Socket } from './net/WebSocketServer'
 import Cache from './services/Cache'
 import { ChatLogRepo } from './services/ChatLogRepo'
